Blueprinted branch competition
