+++ functions_general.py
@@ -1,0 +1,13 @@
+from flask import g
+
+
+def query_db(query, args=(), one=False):
+    cur = g.db.execute(query, args)
+    rv = cur.fetchall()
+    return (rv[0][0] if rv else None) if one else rv
+
+
+def update_db(query, args=()):
+    g.db.execute(query, args)
+    g.db.commit()
+    return
