+++ templates/base.html
@@ -8,7 +8,7 @@
 <body>
     <div id="header">BPL:
         <a href="{{ url_for('home') }}">Home</a> |
-        <a href="{{ url_for('comp') }}">Wettbewerbe</a>
+        <a href="{{ url_for('competition.overview') }}">Wettbewerbe</a>
     </div>
 
     <div id="nav">
