--- templates/comp.html
@@ -1,13 +1,0 @@
-{% extends "base.html" %}
-{%  block content %}
-    <a href="{{ url_for('home') }}"><- Zurück</a>
-    <h1>Wettbewerbe</h1>
-    Wähle einen Wettbewerb
-    <ul>
-    {% for row in table %}
-        <li>
-            <a href='{{  url_for('compdetail', compid=row[0]) }}'>{{ row[1] }}</a>
-        </li>
-    {% endfor %}
-    </ul>
-{% endblock %}
