--- templates/compdetail.html
@@ -1,18 +1,0 @@
-{% extends "base.html" %}
-{%  block content %}
-    <a href="{{ url_for('comp') }}"><- Zurück</a>
-    <h1>Wettbewerb: {{ compname }}</h1>
-    Sieh dir die angemeldeten Teams und Spieler an:
-    <a href="{{ url_for('compteams', compid=compid) }}">Link</a>
-    <br><br>
-    {% if table_unterwb %}
-        Unterwettbewerbe
-        <ul>
-            {% for row in table_unterwb %}
-                <li><a href="{{ url_for('unterwettbewerb', unterwb=row[0]) }}">{{ row[2] }}: {{ row[1] }}</a></li>
-            {% endfor %}
-        </ul>
-    {% endif %}
-    <br><br>
-
-{% endblock %}
