+++ templates/competition.html
@@ -1,0 +1,18 @@
+{% extends "base.html" %}
+{%  block content %}
+    <a href="{{ url_for('competition.overview') }}"><- Zurück</a>
+    <h1>Wettbewerb: {{ compname }}</h1>
+    Sieh dir die angemeldeten Teams und Spieler an:
+    <a href="{{ url_for('competition.competition_signed_teams', competitionid=object[0]) }}">Link</a>
+    <br><br>
+    {% if table_divisions %}
+        Unterwettbewerbe
+        <ul>
+            {% for row in table_divisions %}
+                <li><a href="{{ url_for('competition.division', divisionid=row[0]) }}">{{ row[2] }}: {{ row[1] }}</a></li>
+            {% endfor %}
+        </ul>
+    {% endif %}
+    <br><br>
+
+{% endblock %}
