+++ templates/competition_overview.html
@@ -1,0 +1,13 @@
+{% extends "base.html" %}
+{%  block content %}
+    <a href="{{ url_for('home') }}"><- Zurück</a>
+    <h1>Wettbewerbe</h1>
+    Wähle einen Wettbewerb
+    <ul>
+    {% for row in competition_overview %}
+        <li>
+            <a href='{{  url_for('competition.detail_competition', competitionid=row[0]) }}'>{{ row[1] }}</a>
+        </li>
+    {% endfor %}
+    </ul>
+{% endblock %}
