--- templates/compsite.html
@@ -1,6 +1,0 @@
-{% extends "base.html" %}
-{%  block content %}
-    <h1>This is the site for the competition: {{ name }}</h1>
-    Sieh dir die angemeldeten Teams und Spieler an:
-    <a href="{{ url_for('compteams', compid=compid) }}">Link</a>
-{% endblock %}
