+++ templates/detail_game.html
@@ -1,0 +1,20 @@
+{% extends "base.html" %}
+{%- from 'tablespatterns.html' import kospiele, ligaspiele with context %}
+{%  block content %}
+    <h1>Spiel: {{ gameid }}</h1>
+    Unterwettbewerb: <a href="{{ url_for('unterwettbewerb', unterwb=data[1]) }}">{{ data[2] }}</a><br>
+    {% if modus == 'liga' %}
+        Spieltag: <a href="{{ url_for('league', leagueid=data[1], spieltag=ergebnis[0][7]) }}">{{ ergebnis[0][7] }}</a><br>
+    {%  endif %}
+    Datum: {{ data[3] }}<br>
+    Gewertet: {{ data[4] }}<br>
+    <br>
+    {% if modus == 'liga' %}
+        {{ ligaspiele(ergebnis) }}
+    {% endif %}
+    {% if modus == 'turnier' %}
+    {% endif %}
+    {% if modus == 'ko' %}
+        {{ kospiele(ergebnis) }}
+    {% endif %}
+{% endblock %}
