+++ templates/detail_group.html
@@ -1,4 +1,4 @@
-{% extends "unterwbdetail.html" %}
+{% extends "division.html" %}
 {%  block spiele %}
 
 {% endblock %}
