--- templates/ko.html
@@ -1,6 +1,0 @@
-{% extends "unterwbdetail.html" %}
-{%- from 'tablespatterns.html' import kospiele with context %}
-{%  block spiele %}
-    Spiele:
-    {{ kospiele(table_spiele) }}
-{% endblock %}
