--- templates/league.html
@@ -1,15 +1,0 @@
-{% extends "unterwbdetail.html" %}
-{%- from "tablespatterns.html" import ligaspiele, teamtabelle, spielertabelle with context  %}
-{%  block spiele %}
-    <h1>Spieltag: {{ spieltag }}</h1>
-    <form action="{{ url_for('league', leagueid=unterwbinfo[0], spieltag=spieltag - 1) }}">
-        <input type="submit" value="Spieltag zurück" {% if spieltag == 1 %}disabled{% endif %}>
-    </form> <form action="{{ url_for('league', leagueid=unterwbinfo[0], spieltag=spieltag + 1) }}">
-        <input type="submit" value="Spieltag weiter" {% if spieltag == max_spieltag %}disabled{% endif %}>
-    </form>
-    {{ teamtabelle(table_teamtabelle) }}
-    <br><br>
-    {{ ligaspiele(table_spiele) }}
-    <br><br>
-    {{ spielertabelle(table_spielertabelle) }}
-{% endblock %}
