--- templates/unterwbdetail.html
@@ -1,10 +1,0 @@
-{% extends "base.html" %}
-{% block content %}
-    <a href="{{ url_for('compdetail', compid=unterwbinfo[3]) }}"><- Zurück</a>
-    <h1>Unterwettbewerb: {{ unterwbinfo[1] }}</h1>
-    Sieh dir die angemeldeten Teams und Spieler an:
-    <a href="{{ url_for('unterwbteams', unterwb=unterwbinfo[0]) }}">Link</a>
-    <br>
-    <br>
-    {% block spiele %}{% endblock %}
-{% endblock %}
